requirements.txt
PyQt6==6.7.0
Pillow==10.2.0